v0.5.0 pre-tag Group B (S3): relocate *Client interfaces from core to redis
Pre-tag interface review S3: backing-client interfaces (ChallengeStoreClient,
ReplaySeenSetClient, RefreshTokenFamilyClient/Multi/Disposable,
UserSessionStoreClient, SessionRPRegistryClient/Multi,
SessionSidSortedSetClient/Multi, FederationTokenStoreClient,
RateLimiterClient) are intrinsically Redis-flavoured (hSet, zAdd, pttl,
multi/watch/exec, pExpireAt). Hosting them in @o3co/auth-provider-core forced
storage-agnostic consumers to mimic Redis semantics.

Move:
- 11 type declarations + 9 ComponentMap slot augmentations relocate from
  core type files to a new packages/redis/src/clients.mts. ComponentMap
  augmentations still target @o3co/auth-provider-core via declare module so
  the slot keys remain unchanged at the type level; only the augmentation
  source moves.
- packages/core/src/index.mts no longer exports any *Client type.
- packages/redis/src/index.mts exports the 11 *Client types directly from
  ./clients.mjs (no more re-export from core).
- All 16 redis source/test files that previously imported *Client from core
  now import from local ./clients.mjs (or ../src/clients.mjs from tests).
- biome import sort applied.

Custom non-Redis backends (DynamoDB, Postgres, etcd, ...) define their own
client contracts; do NOT implement these Redis-shaped interfaces.

Validation: 2136 tests pass, typecheck clean across all 10 packages,
lint clean (15 pre-existing warnings unaffected). Initial flaky failure
in oauth (supertest port parse error) resolved on retry — not introduced
by this commit.

Closes #108
