+++ packages/core/src/ratelimit/types.mts
@@ -69,19 +69,5 @@ declare module "@o3co/auth-provider-core" {
 // Backing client interface (Phase 10 addendum §3)
 // ---------------------------------------------------------------------------
 
-/**
- * Backing client for RateLimiter adapters. Declares only `incr` and `expire`
- * — the two methods `createRedisRateLimiter` consumes.
- *
- * Per Phase 10 addendum §3.
- */
-export interface RateLimiterClient {
-	incr(key: string): Promise<number>;
-	expire(key: string, seconds: number): Promise<number>;
-}
-
-declare module "@o3co/auth-provider-core" {
-	interface ComponentMap {
-		readonly rateLimiterClient?: RateLimiterClient;
-	}
-}
+// RateLimiterClient backing-client interface relocated to
+// @o3co/auth-provider-redis (v0.5.0 pre-tag interface review S3).
