+++ packages/core/src/user-sessions/types.mts
@@ -185,84 +185,6 @@ declare module "@o3co/auth-provider-core" {
 // Backing client interfaces (Phase 10 addendum §3)
 // ---------------------------------------------------------------------------
 
-/**
- * Backing client for UserSessionStore adapters. Declares only `set`, `get`,
- * `del` — the exact methods `createRedisUserSessionStore` consumes.
- * The `set` overloads cover both the plain-PX form (for future update paths)
- * and the PX+NX form used by `create` (atomic insert-only, mirrors
- * ChallengeStore.issue and RefreshTokenFamilyStore.registerFamily).
- *
- * Per Phase 10 addendum §3 (NX overload added based on actual adapter audit
- * of `userSessionStore.mts`; plan table erroneously described "no NX").
- */
-export interface UserSessionStoreClient {
-	set(key: string, value: string, mode: "PX", ttlMs: number): Promise<"OK" | null>;
-	set(key: string, value: string, mode: "PX", ttlMs: number, condition: "NX"): Promise<"OK" | null>;
-	get(key: string): Promise<string | null>;
-	del(key: string): Promise<number>;
-}
-
-/**
- * Chainable transaction pipeline returned by `SessionRPRegistryClient.multi()`.
- *
- * Per Phase 10 addendum §3.
- */
-export interface SessionRPRegistryMultiClient {
-	hSet(key: string, field: string, value: string): SessionRPRegistryMultiClient;
-	pExpireAt(key: string, msTimestamp: number): SessionRPRegistryMultiClient;
-	exec(): Promise<unknown[] | null>;
-}
-
-/**
- * Backing client for SessionRPRegistry adapters. Declares the hash ops +
- * multi pipeline that `createRedisSidHash` consumes.
- *
- * Per Phase 10 addendum §3.
- */
-export interface SessionRPRegistryClient {
-	del(key: string): Promise<number>;
-	hSet(key: string, field: string, value: string): Promise<number>;
-	hVals(key: string): Promise<string[]>;
-	multi(): SessionRPRegistryMultiClient;
-	pExpireAt(key: string, msTimestamp: number): Promise<number>;
-}
-
-/**
- * Chainable transaction pipeline returned by `SessionSidSortedSetClient.multi()`.
- *
- * Per Phase 10 addendum §3.
- */
-export interface SessionSidSortedSetMultiClient {
-	pExpireAt(key: string, msTimestamp: number): SessionSidSortedSetMultiClient;
-	zAdd(
-		key: string,
-		entry: { score: number; value: string },
-		opts?: { NX: true },
-	): SessionSidSortedSetMultiClient;
-	exec(): Promise<unknown[] | null>;
-}
-
-/**
- * Backing client for SessionFamilyIndex and SessionFederationIndex adapters.
- * Both adapters share this interface (same sorted-set operations, different
- * slot identities in ComponentMap).
- *
- * Per Phase 10 addendum §3.
- */
-export interface SessionSidSortedSetClient {
-	del(key: string): Promise<number>;
-	multi(): SessionSidSortedSetMultiClient;
-	pExpireAt(key: string, msTimestamp: number): Promise<number>;
-	zAdd(key: string, entry: { score: number; value: string }, opts?: { NX: true }): Promise<number>;
-	zRange(key: string, start: number, stop: number): Promise<string[]>;
-	zRem(key: string, member: string): Promise<number>;
-}
-
-declare module "@o3co/auth-provider-core" {
-	interface ComponentMap {
-		readonly userSessionStoreClient?: UserSessionStoreClient;
-		readonly sessionRPRegistryClient?: SessionRPRegistryClient;
-		readonly sessionFamilyIndexClient?: SessionSidSortedSetClient;
-		readonly sessionFederationIndexClient?: SessionSidSortedSetClient;
-	}
-}
+// UserSessionStoreClient / SessionRPRegistryClient (+Multi) /
+// SessionSidSortedSetClient (+Multi) backing-client interfaces relocated to
+// @o3co/auth-provider-redis (v0.5.0 pre-tag interface review S3).
